Remove Dialogue Autoplay and NPCAutoFollow classes and combined functionality into dialogueInteractable and NPCInteractable
