+++ Assets/Scripts/Interactables/Interactable.cs
@@ -16,7 +16,7 @@ public class Interactable : MonoBehaviour, IInteractable
     public delegate void OnInteractEventHandler(Interactable source);
     public event OnInteractEventHandler OnInteractEnd;
 
-    protected void Update()
+    protected virtual void Update()
     {
         //Ensures text is always facing the camera
         if(interactableOn)
